requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
twilio==7.8.0
Pillow>=9.2.0
pytesseract==0.3.8
opencv-python-headless>=4.5.3.56
redis==3.5.3
python-dotenv==0.19.0
APScheduler==3.8.0
gunicorn==20.1.0